requirement.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-toolbelt==0.0.1
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
static3==0.7.0
urllib3==1.23
whitenoise==4.1
